Rename to simulated_annealing_generator
