--- modules/generate/src/simulated_annealing.cpp
@@ -1,1 +1,0 @@
-#include "simulated_annealing.hpp"
