+++ modules/generate/src/simulated_annealing_generator.cpp
@@ -1,0 +1,1 @@
+#include "simulated_annealing_generator.hpp"
